feat(stats): collect resource stats during a run's start phase
Builds, installs and dev-server boot-up were the one part of a run nothing
observed. The daemon's sampler skipped any run that was not RunStatus::Running,
and a run is Starting from before its first stage until every node is healthy —
so the entire start phase produced no samples. Worse, a command step's PID is
created, awaited and reaped inside the CLI, so nothing outside that process
could ever see a build at all.

Two changes, and one invariant that keeps them apart:

- The daemon now samples Starting runs as well as Running ones, in both the
  sampler and /api/stats. A start_server node checkpoints its PID before its
  health check, so this alone records the whole boot-up allocation ramp.
- The CLI samples the command steps it runs, every 2s, via a new
  veld_stats::CommandStatsRecorder installed on the orchestrator through the
  sysinfo-free veld_core::stats::StepObserver seam.

The transient PID is measured, never persisted. NodeState.pid is a claim that a
node has a process now — read by veld stop, the health monitor, the GC, the
registry's has-ever-spawned query and is_reapable_orphan — so a finished build
stored there would make a run that is still legitimately coming up look like one
that spawned and died, and would eventually have veld stop signal whatever
recycled the number. A sample is the opposite kind of fact and stays true after
the process is gone.

Sampling moves to its own crate so the two producers cannot drift in what a
number means, and so veld-helper (privileged) and veld-gateway — both of which
depend on veld-core — do not link a machine-wide process scanner.

No migration, no schema change, no config surface.
