+++ AGENTS.md
@@ -259,6 +259,27 @@ and several were paid for in this codebase already.
   `pull_request` only — a superseded PR run is waste, but a half-cancelled push
   to `main` is a release that tagged without publishing.
 - **Any user-supplied command executed by a daemon must inherit the user's login-shell `PATH`.** The daemon (launchd), gateway (systemd), and helper run with a bare service `PATH`, so a raw `sh -c` cannot find user-installed CLIs (`op`, `vault`, `pg_isready`, version-manager shims) even though the same command works in the user's terminal. Pass it via `.env("PATH", …)`, resolved by one of the two helpers in `veld_core::user_path`. Anything inside a daemon uses `cached_user_path()` — every request handler (`spawn_veld`, the desktop directory picker and git plumbing, and `SecretSource::Command` token resolution in `veld-share/src/endpoint.rs`, which `POST /api/shares` reaches while the caller waits) and the health monitor's liveness probes; a dedicated `warm_user_path_cache` task keeps that entry warm, deliberately **not** the monitor's scan, whose own cadence stretches to 300s during a recovery restart. `resolve_user_path()` is the uncached primitive, for a one-shot context that resolves once and exits — today only a CLI run's lazy var sources. Resolution spawns a login shell: sub-second normally, up to 10s on a stalled rc file, and the slow rc files belong to exactly the users who need the resolved PATH — which is why a handler must not resolve inline. The cache is one writer (the warm task) publishing into a cell that readers only read, with a single rule doing the work: **a resolution that learned nothing never displaces one that did.** A stall, or a shell that exits zero while answering with nothing better than the process's own `PATH`, leaves the previous value alone — because that value is the bare service PATH, i.e. the bug itself. Resist adding a TTL or a which-resolution-wins rule back: the first version of this cache had both, and each grew its own bug. Never spawn a config-declared command on a daemon without this. **A login shell is not a substitute for the injection.** `$SHELL -l -c` sources `.zprofile` but not `.zshrc`, which is where version managers (nvm, fnm, rbenv) and `brew shellenv` live on most machines; on Debian `/etc/profile` overwrites `PATH` outright, so a login shell wrapped around an injected PATH *discards* it. That mistake shipped in `spawn_veld` and made every UI- or Desktop-initiated `veld start` resolve node commands against the bare launchd PATH (`sh: npx: command not found`). Spawn the binary directly with the injected PATH; don't reach for a shell to get it. **The PTY holder is the documented exception, and now has an exception of its own.** `spawn_shell` skips the helper because the thing it spawns *is* a login shell, which computes the same `PATH` itself. But a config-declared pane (`ide.panes`) is spawned as an `argv` with **no login shell in front of it**, so the rule applies again in full: the daemon resolves `cached_user_path()` at ticket-mint time and passes it in `HolderConfig::env`. Get that wrong and every pane fails with `command not found` for the exact CLI it exists to run, while a plain terminal in the same app works perfectly — `a_pane_command_runs_instead_of_a_shell_and_gets_the_injected_path` pins it with a real process. Scope: the rule covers daemon/gateway/helper spawns only — commands the `veld` CLI itself spawns (orchestrator `command`/`start_server` steps, setup checks, actions) already inherit the terminal's `PATH` and are exempt. Only `PATH` is inherited, never the rest of the shell environment — and note that this genuinely differs from the CLI path, where a macOS terminal's zsh *is* a login shell, so `.zprofile` exports (`JAVA_HOME`, `LANG`, tool tokens) do reach a terminal-run `veld` and its node commands but not a daemon-spawned one. A node that needs such a variable must declare it in its `env`; "works in my terminal, fails from the UI" for a non-PATH variable is this asymmetry, not a bug.
+- **A transient PID is measured, never persisted — and the two stats producers
+  stay disjoint by node kind.** Resource samples come from two places:
+  `veld-daemon` samples every node with a persisted `NodeState.pid`, and the
+  `veld` CLI samples the `command` steps it runs (`veld_stats::CommandStatsRecorder`,
+  installed via `Orchestrator::with_step_observer`) because a build's process is
+  spawned, awaited and reaped inside that command and its PID exists nowhere
+  else. Both call the same `veld_stats::StatsCollector`, so the two producers
+  cannot drift in what a number means. The obvious "simplification" — persist the
+  `command` step's PID so the daemon can sample it like everything else — is the
+  one thing that must not happen: `NodeState.pid` is a claim that the node has a
+  process *now*, read by `veld stop`, the health monitor, the GC, the registry's
+  "has ever spawned" query and `is_reapable_orphan`, so a finished build sitting
+  there makes a run that is still legitimately coming up look like one that
+  spawned and died, and eventually has `veld stop` signal whatever recycled the
+  number. A sample is the opposite kind of fact — "at time T this tree looked like
+  this" — and stays true afterwards. `only_the_start_server_path_ever_persists_a_node_pid`
+  pins it by counting the assignment sites, because nothing in the type system
+  does. The sampling code lives in its own crate rather than in `veld-core` for a
+  second reason: `veld-helper` (privileged) and `veld-gateway` depend on
+  `veld-core`, and neither should link a machine-wide process scanner — hence the
+  sysinfo-free `veld_core::stats::StepObserver` seam that the CLI implements.
 - **`rg` needs `-a` on `crates/veld-daemon/ui/src/App.tsx`.** That file contains a
   NUL byte (`trustedOrigins.join("\0")`), so ripgrep and grep classify the largest
   file in the UI as binary and **silently skip it** — no match, no warning. A sweep
