+++ README.md
@@ -39,7 +39,7 @@ No port numbers. No manual wiring. Just clean, stable, human-readable URLs.
 - **Every step's output is collected** — `command` nodes (a `docker build`, a `pnpm install`) log to their node's stream exactly like servers do, and project `setup`/`teardown` steps to the run's; nothing a step prints is thrown away. Lines also stream into `veld start`'s progress output as they arrive, instead of scribbling over it
 - **Client-side logs** — captures browser `console.log/warn/error`, exceptions, and promise rejections; view with `veld logs --source client`
 - **Internal logs** — liveness probe outcomes (with stderr), recovery decisions, health state transitions; view with `veld logs --source internal`
-- **Resource monitoring that doesn't lie** — the daemon samples every node's whole process tree every 5s. The headline memory figure is the tree's **footprint** (proportional set size on Linux, `phys_footprint` on macOS), because summing RSS over a tree counts each page shared inside it once per process, so a five-process `npm run dev` reported far more than it occupied. On Linux the footprint splits by page class — private dirty (the heap, what grows when a node leaks), private clean, shared dirty, shared clean, swap, wired — and every node splits by **subprocess**, so "which child is eating the RAM" is a question with an answer. `veld stats` shows it in the terminal (`--processes`, `--history`, `--cpu`, `--memory <metric>`); the dashboard expands any node into a scrubbable chart you can flip between memory and CPU, and between total, by-type and by-process. Where a bucket averages several samples, the chart plots the peak alongside the mean — a mean over a six-minute bucket hides the spike a 5s sample caught. Totals are kept 24h, per-process detail 2h (both reported by the API, so nothing hardcodes them); `VELD_STATS_MEMORY_DETAIL=off` and `VELD_STATS_CMDLINE=off` turn off the detailed probe and argv capture respectively — set in the *daemon's* environment (`launchctl setenv` / `systemctl --user set-environment`, then restart it; a shell `export` does not reach a launchd/systemd service), see [skills/veld/SKILL.md](skills/veld/SKILL.md)
+- **Resource monitoring that doesn't lie, including while a run is still starting** — the daemon samples every node's whole process tree every 5s, from the moment the run begins rather than once it is fully up: a dev server's boot-up allocation ramp is recorded, not skipped. `command` steps — builds, installs, codegen — are sampled too, every 2s, by the `veld start` that runs them, because their processes never outlive that command and no PID for them exists anywhere else. So "what did that `cargo build` peak at" and "how much does the dev server hold before it serves its first request" are questions with answers. (A `docker build` is the exception: the work happens inside `dockerd`/`buildkitd`, outside the step's process tree, so the client is all veld can see.) The headline memory figure is the tree's **footprint** (proportional set size on Linux, `phys_footprint` on macOS), because summing RSS over a tree counts each page shared inside it once per process, so a five-process `npm run dev` reported far more than it occupied. On Linux the footprint splits by page class — private dirty (the heap, what grows when a node leaks), private clean, shared dirty, shared clean, swap, wired — and every node splits by **subprocess**, so "which child is eating the RAM" is a question with an answer. `veld stats` shows it in the terminal (`--processes`, `--history`, `--cpu`, `--memory <metric>`); the dashboard expands any node into a scrubbable chart you can flip between memory and CPU, and between total, by-type and by-process. Where a bucket averages several samples, the chart plots the peak alongside the mean — a mean over a six-minute bucket hides the spike a 5s sample caught. Totals are kept 24h, per-process detail 2h (both reported by the API, so nothing hardcodes them); `VELD_STATS_MEMORY_DETAIL=off` and `VELD_STATS_CMDLINE=off` turn off the detailed probe and argv capture respectively — set in the *daemon's* environment (`launchctl setenv` / `systemctl --user set-environment`, then restart it; a shell `export` does not reach a launchd/systemd service), see [skills/veld/SKILL.md](skills/veld/SKILL.md)
 - **Reverse-proxy header rules** — add or strip request/response headers on the local proxy and the public web gateway with a `proxy` config block (project/node/variant). Veld does no header manipulation by default.
 - **Peer-to-peer sharing** — share a running environment with a colleague over an encrypted P2P tunnel (`veld share`); they open the same URLs on their own machine. Services opt in explicitly in config, and relays are configurable (public or self-hosted) for compliance. No accounts, no Veld-hosted server.
 - **Public web sharing** — expose a service to someone *without* Veld (`veld share --web`): a self-hosted gateway (`veld-gateway`, one Docker container) mints a real public URL anyone can open in a browser. The overlay's **Copy public URL** action translates your current page (path + query preserved) into the public link.
