+++ crates/veld-daemon/Cargo.toml
@@ -13,6 +13,7 @@ path = "src/main.rs"
 [dependencies]
 veld-core = { workspace = true }
 veld-share = { workspace = true }
+veld-stats = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true }
@@ -23,7 +24,6 @@ tracing-subscriber = { workspace = true }
 chrono = { workspace = true }
 dirs = { workspace = true }
 libc = "0.2"
-sysinfo = { version = "0.33", default-features = false, features = ["system"] }
 axum = { version = "0.8", features = ["ws"] }
 reqwest = { workspace = true }
 uuid = { workspace = true }
