+++ crates/veld-stats/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "veld-stats"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+rust-version.workspace = true
+publish.workspace = true
+
+[dependencies]
+veld-core = { workspace = true }
+sysinfo = { version = "0.33", default-features = false, features = ["system"] }
+chrono = { workspace = true }
+tokio = { workspace = true }
+tracing = { workspace = true }
+libc = "0.2"
+
+[dev-dependencies]
+tempfile = "3"
