+++ website/llms-full.txt
@@ -281,6 +281,8 @@ Each run card's Services tab also shows live resource usage per node — memory 
 
 ### Resource measurement
 
+Sampling covers a run's **start phase**, not only its steady state, and which process does the sampling depends on the step. Long-lived services are sampled by the daemon from the moment they spawn — the whole boot-up window before the service is healthy, which is where a dev server does most of its allocating. Build and install steps are sampled every ~2s by the `veld start` that runs them, because such a step's process is spawned, awaited and reaped inside that command and its process id exists nowhere else; publishing that id as service state instead would make a finished build look, to `veld stop` and to the orphan reaper, like a service that died. Two consequences when reading the data: a build has no *live* reading once it finishes — its curve lives in the history, which is where a peak is read from — and a step shorter than one sampling interval is represented by the single sample taken when it spawned. A `docker build` is the exception that stays invisible: the work happens inside `dockerd`/`buildkitd`, which are not descendants of the step, so the client is all there is to measure.
+
 The memory figure Veld reports is the process tree's **footprint**, not the sum of its processes' resident set sizes. Summing RSS counts every page shared *inside* the tree once per process — the shared libc, the shared Node runtime, every copy-on-write page inherited at fork — so a five-process `npm run dev` reports far more than it occupies. Footprint is proportional set size on Linux (each shared page divided by the number of processes mapping it, which makes a tree sum honest) and `phys_footprint` on macOS (the number Activity Monitor shows).
 
 On Linux the footprint also splits by page class, read from `/proc/<pid>/smaps_rollup`: private dirty (the heap — what grows when a service leaks), private clean, shared dirty, shared clean, swap, and locked. macOS has no unprivileged equivalent, so it reports the totals plus wired memory and says so rather than inventing a split; a metric that cannot be measured is reported as absent, never as zero.
